// src/api/generated/endpoints/lessons-api.ts
export interface LessonsApiApproveLessonRequest {
    /**
     * 
     * @type {string}
     * @memberof LessonsApiApproveLesson
     */
    readonly lessonUnitId: string
}

/**
 * Request parameters for createLessonTopic operation in LessonsApi.
 * @export
 * @interface LessonsApiCreateLessonTopicRequest
 */
export interface LessonsApiCreateLessonTopicRequest {
    /**
     * 
     * @type {CreateLessonTopicDto}
     * @memberof LessonsApiCreateLessonTopic
     */
    readonly createLessonTopicDto?: CreateLessonTopicDto
}

/**
 * Request parameters for getLesson operation in LessonsApi.
 * @export
 * @interface LessonsApiGetLessonRequest
 */
export interface LessonsApiGetLessonRequest {
    /**
     * 
     * @type {string}
     * @memberof LessonsApiGetLesson
     */
    readonly lessonUnitId: string
}

/**
 * Request parameters for getLessonTopics operation in LessonsApi.
 * @export
 * @interface LessonsApiGetLessonTopicsRequest
 */
export interface LessonsApiGetLessonTopicsRequest {
    /**
     * 
     * @type {number}
     * @memberof LessonsApiGetLessonTopics
     */
    readonly termId?: number

    /**
     * 
     * @type {string}
     * @memberof LessonsApiGetLessonTopics
     */
    readonly classId?: string

    /**
     * 
     * @type {string}
     * @memberof LessonsApiGetLessonTopics
     */
    readonly subjectId?: string
}

/**
 * LessonsApi - object-oriented interface
 * @export
 * @class LessonsApi
 * @extends {BaseAPI}
 */
export class LessonsApi extends BaseAPI {
    /**
     * 
     * @param {LessonsApiApproveLessonRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof LessonsApi
     */
    public approveLesson(requestParameters: LessonsApiApproveLessonRequest, options?: RawAxiosRequestConfig) {
        return LessonsApiFp(this.configuration).approveLesson(requestParameters.lessonUnitId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {LessonsApiCreateLessonTopicRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof LessonsApi
     */
    public createLessonTopic(requestParameters: LessonsApiCreateLessonTopicRequest = {}, options?: RawAxiosRequestConfig) {
        return LessonsApiFp(this.configuration).createLessonTopic(requestParameters.createLessonTopicDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {LessonsApiGetLessonRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof LessonsApi
     */
    public getLesson(requestParameters: LessonsApiGetLessonRequest, options?: RawAxiosRequestConfig) {
        return LessonsApiFp(this.configuration).getLesson(requestParameters.lessonUnitId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {LessonsApiGetLessonTopicsRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof LessonsApi
     */
    public getLessonTopics(requestParameters: LessonsApiGetLessonTopicsRequest = {}, options?: RawAxiosRequestConfig) {
        return LessonsApiFp(this.configuration).getLessonTopics(requestParameters.termId, requestParameters.classId, requestParameters.subjectId, options).then((request) => request(this.axios, this.basePath));
    }
}